changing the code 3 to code iso in country to reflect better the business model adding person service to the country service and message source to the country handler changing the code  to code  iso in country dao deleting useless method from country repository deleting useless test from country repository test changing the code 3 to code  iso remove unneeded flowof in country handler adding message source to country handler deleting the sql from the country repository deleting the unneeded method in country repository adding annotation to country router for  swagger ui creating country router info for swagger annotation fixing country router test adding remover method from the repository to the service adding tests in country service test changing the FraudsterAndCountryCodeDTO to data class moving the sonar params to the maven yml file adding more messages to messages.properties deleting the @Autowired annotation removing the mono and refactoring the code of PersonDatabaseRepository.kt adding new tests to the PersonDatabaseRepositoryTest.kt changing the PersonDTO.kt to data class and removing the builder name adding missing errors to the PersonHandler.kt creating PersonInCountry.kt data class changing the PersonsByCountry.kt number of user to int refactoring person service adding specific errors linting PersonServiceTest.kt removing sonar config from pom changing the code3 to code3 iso in schema.sql adding OpenAPIDefinition to main class

+++ src/main/kotlin/tn/keyrus/pfe/imdznd/userservice/dirtyworld/person/dao/PersonsByCountry.kt
@@ -2,5 +2,5 @@ package tn.keyrus.pfe.imdznd.userservice.dirtyworld.person.dao
 
 data class PersonsByCountry(
     val country: String = "",
-    val numberOfUsers: Long = 0,
+    val numberOfUsers: Int = 0,
 )
